Fix /v1/cards
The documentation seems to incorrectly include withBalance as an optional
boolean query parameter. It also, like other methods includes the consent-id
header param that is  not in the YAML spec. The documentation however
specifies that the return is 'an array of Card details' and further
elaborates on that 'some attributes properties of the
dedicated card are provided' instead of just the account
information It therefore makes no sense when the
readCardListResponse-200_json type then returns cardAccontDetails
instead of cardDetails. Actually that type is not used as such
in the contract. This seems to warrant a change to the return
and the example as well, which only specifies cardAccount details.
Defenition for bank was missing and set as code with 3-4 numbers.