Fix links in card detail.
The GET /v1/cards/{card-id} does not include direct errors,
except the usual C&P withBalance query param and missing
consent-it. However the document text is inconsistent, seems
to indicate the account info should be returned but specifies
the type as card detail. It is suggested that this should be the
card detail, used when the user/TPP has stored the ID for later
retrieval.
The account type includes a _links type copied from the PSD2 schema
but which is inconsistent with the example and general practice,
in not defining the "self" link. All this is also currently missing
from the detail type, which should be fixed so e.g. the list is
more helpful.
PanId was missing description.