Fix /v1/cards/{card-id}/balances
Besides the usual inconsistencies, the documentation does not
mention the dateFrom query param that the YAML indicates
should retrieve balances by date. On the whole, this param
makes sense but needs description.
At this point in time it might be noted that documentation for this
method notes limitations to the current line of credit being possibly
more constrained than the balance reflects, in the document and
YAML. As other returns include balance, that should apply elsewhere.