Fix /v1/cards/{card-id}/transactions:
The summary and description in the YAML was incorrect. The documentation
sticks to being incorrect on withBalances, but skips describing the delta
list parameter that would warrant further defenition. The YAML and the
documentation then start to disagree on more important points in
readCardTransactionsResponse-200_json:
card:  cardAccountDetails instead of accountreference
cardTransactions: missing in YAML, should be cardAccountReport.
cardAccountReport then seems to support paging in YAML but
documentation only supports download of a transaction report when
containing a large amount of data. Which was desired back in the day
FUT only knows, but the download seems to have been discussed and
with all the other mismatches in the YAML, that is chosen here.
Some link inconsistencies cleaned up.