Fix Berlin Group NG reference overload
The Cards schema is not part of BG NG and has, after being
normalized no direct references. Nor is IOBWS exclusvively
meant for TPP consumption, though that concern remains.
Therefore some of the most glaring BG NG referneces, extraneous
types and functionality is removed in this iteration.
SVC is used instead of TPP, for intermediary service which
has a broader reference. User replaces PSU.
Still proper attribution is called for and added to the header.