Fix currencies
/v1/currencies
The root currencies method has only minor inconsistencies
in the TS text, but the YAML references 'cards' in the response. The
response was then missing an example.

/v1/currencies/sources
Mostly correct, however, having currencySourceQuery
as part of the query filter for /sources, which retrieves a type with a
single field is not according to documentation, or common sense.
Missing an example.

/v1/currencies/{base-currency}/rates:
The text for this and other methods is somewhat muddled
as the term 'single currency' is used for either a base or
quote currency. Tried fixing. The return type OK_200_CURRENCY_RATES
for this method was not in accordance with naming conventions
and was renamed to OK_200_BASE_CURRENCY_RATES. Added missing
example.

/v1/currencies/{quote-currency}/rates/{base-currency}:
Here the operation ID could be debated when looking at consistency
and the suggestion would be to refer to the cross against a
quoted currency, as well as having the methods change signature
to use the standard base/quoted order. Not done, but
added missing example.

/v1/currencies/{quote-currency}/rates/{base-currency}/history:
This and the previous method switch the order of base/quote and
the documentation refers to cards, while an example was added.
dateFrom and dateTo referenced accounts/cards and the text was changed
to be more general.