Proposed changes to Digital Documents (ÍST TS 314:2022) (#229)
* Proposed changes to IOBWS-Documents3.0.yaml

These are proposed changes to IOBWS-Documents3.0.yaml due to cross-references. 

For further information, see corresponding issue:

https://github.com/stadlar/IST-FUT-FMTH/issues/227

* minor updates on ÍST 314 description and service definition

The PR was merged into the feature branch that will be used for the workgroup's efforts, with the possibility of accepting further PRs from participants and at least some minor updates to the documentations relevant for the update to the Technical Standard.