Proposed changes to Digital Documents (ÍST TS 314:2022) (#229) (#230)
Merged based on approval in meeting TN FMÞ VH7-31 on the 19th of March. Included:

* Proposed changes to Digital Documents (ÍST TS 314:2022) (#229)

* Proposed changes to IOBWS-Documents3.0.yaml

These are proposed changes to IOBWS-Documents3.0.yaml due to cross-references. 

For further information, see corresponding issue:

https://github.com/stadlar/IST-FUT-FMTH/issues/227

* minor updates on ÍST 314 description and service definition

The PR was merged into the feature branch that will be used for the workgroup's efforts, with the possibility of accepting further PRs from participants and at least some minor updates to the documentations relevant for the update to the Technical Standard.

* Document updated for 314 and PDF generated

* Update CHANGELOG for TS 314 point release and document validation

* Update CHANGELOG for TS 314 point release and implement proposed changes from #227

* Update document variables and metadata for ÍST TS 314:2025, including date adjustments and participant details

* Update funding paragraph in docvariables.yml 

* Update rending of  ÍST TS 314:2025 document

---------

Co-authored-by: andrrri <gudm.andri.hjalmarsson@rb.is>